requirements.txt
beautifulsoup4==4.13.4
chromadb==0.5.5
langchain==0.3.25
langchain_community==0.3.24
matplotlib==3.10.3
numpy==2.2.6
openai==1.79.0
pandas==2.2.3
pypdf==5.5.0
python-dotenv==1.1.0
Requests==2.32.3
scikit_learn==1.6.1
sentence_transformers==4.1.0
streamlit==1.45.0
torch==2.7.0
transformers==4.51.3
umap==0.1.1
umap_learn==0.5.7
